Code refactor and cleanup.
